console: Address small remarks, make separate container component
