+++ pkg/webui/locales/en.json
@@ -377,6 +377,9 @@
   "console.containers.gateways-table.index.restoreFail": "There was an error and the gateway could not be restored",
   "console.containers.gateways-table.index.purgeSuccess": "Gateway purged",
   "console.containers.gateways-table.index.purgeFail": "There was an error and the gateway could not be purged",
+  "console.containers.log-back-in-modal.index.modalTitle": "Please sign in again",
+  "console.containers.log-back-in-modal.index.modalMessage": "You were signed out of the Console. You can press 'Reload' to log back into the Console again.",
+  "console.containers.log-back-in-modal.index.buttonMessage": "Reload",
   "console.containers.lora-cloud-das-form.index.tokenDescription": "Device & Application Services Access Token as configured within LoRa Cloud",
   "console.containers.lora-cloud-das-form.index.lr1110Encoding": "LR1110 demo encoding",
   "console.containers.lora-cloud-das-form.index.setLoRaCloudToken": "Set LoRa Cloud token",
@@ -517,9 +520,6 @@
   "console.views.admin-packet-broker.messages.Are you sure you want to deregister your network from Packet Broker?{lineBreak}This will <b>permanently delete<.b> all routing policies and may stop traffic from flowing.{lineBreak}Traffic may still be forwarded to your network based on default routing policies configured by forwarders.": "Are you sure you want to deregister your network from Packet Broker?{lineBreak}This will <b>permanently delete</b> all routing policies and may stop traffic from flowing.{lineBreak}Traffic may still be forwarded to your network based on default routing policies configured by forwarders.",
   "console.views.admin-user-management-edit.index.updateSuccess": "User updated",
   "console.views.admin-user-management-edit.index.deleteSuccess": "User deleted",
-  "console.views.app.index.modalTitle": "Please sign in again",
-  "console.views.app.index.modalMessage": "You were signed out of your account. Please press 'Reload' to log back in the console again.",
-  "console.views.app.index.buttonMessage": "Reload",
   "console.views.application-add.index.applicationName": "Application name",
   "console.views.application-add.index.appIdPlaceholder": "my-new-application",
   "console.views.application-add.index.appNamePlaceholder": "My new application",
